refactor: move location of functionality of dat parsing
